api/requirements-dev.txt
flake8==3.6.0
pytest-cov==2.6.0
pytest-tornado==0.5.0
pytest==3.10.1
wheel==0.32.2
